requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
openai==1.12.0
pdfplumber==0.10.3
PyMuPDF==1.23.8
pandas==2.1.4
numpy==1.24.3
faiss-cpu==1.7.4
stripe==7.8.0
werkzeug==3.0.1
sqlalchemy==2.0.23
bcrypt==4.1.2
requests==2.31.0
langchain==0.1.0
tiktoken==0.5.2
tesseract==0.1.3
pytesseract==0.3.10
Pillow==10.1.0 